double changes

+++ VTK/Hybrid/vtkXYPlotActor.cxx
@@ -1793,13 +1793,13 @@ int vtkXYPlotActor::GetPlotPoints(int i)
 }
 
 //----------------------------------------------------------------------------
-void vtkXYPlotActor::SetPlotColor(int i, float r, float g, float b)
+void vtkXYPlotActor::SetPlotColor(int i, double r, double g, double b)
 {
   this->LegendActor->SetEntryColor(i, r, g, b);
 }
 
 //----------------------------------------------------------------------------
-float *vtkXYPlotActor::GetPlotColor(int i)
+double *vtkXYPlotActor::GetPlotColor(int i)
 {
   return this->LegendActor->GetEntryColor(i);
 }
